Add to sample an interface that encapsulates HTTP traffic
