+++ sample/src/main/java/com/chuckerteam/chucker/sample/HttpTask.kt
@@ -1,0 +1,5 @@
+package com.chuckerteam.chucker.sample
+
+interface HttpTask {
+    fun run()
+}
